list

+++ course_from_scratch/first_steps/math/rectangle_area.py
@@ -1,7 +1,6 @@
 print("enter the area of the rectangle")
-pole1 = int(input("Enter pole 1: "))
-pole2 = int(input("Enter pole 2: "))
+pole1 = float(input("Enter pole 1: "))
+pole2 = float(input("Enter pole 2: "))
 
 pole3 = pole1 * pole2
 print(pole3)
-
